refactor/inline event-clock to fairqueuing package
